+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change log
 
+## 0.15.1 (2024-10-XX)
+
+### Features
+
+- popup: Show visited date as related time value
+- popup: Improved visited releases widget
+
 ## 0.15.0 (2024-09-30)
 
 ### Features